Added a PoseStamped message to the atomic_imu node.
This allows us to visualize what the IMU is sending as far as orientation.  Position is hard-coded to (0,0,0).

Also, we found out what the error message with the robot_pose_ekf filter is.  the robot_pose_ekf filter is essentially telling us that two of the sensors going into the filter are varying wildly.  In our case, it was the only two sensors enabled in the launch file, the wheel odometry and the IMU.  It is my estimation that the IMU is severely off somewhere, so Caleb is going to take a closer look at it.