Added magnetic direction and declination for GPS.
The GPS coordinate frame is now instantiated by the magnetic_direction and magnetic_declination parameters set in the dg14_gps.launch file

The magnetic_direction variable is the compass heading of the robot when it is turned on.

The magnetic_declination variable is the magnetic declination for the area where the robot will be operating, since GPS uses
true north rather than magnetic north.  The declination should be postive for eastern declination and negative for western declination.

Calvin's House == -3.5 degrees (West)
Beaver Creek == -6 degrees (West)